공격력 스탯이 조금 이상함

+++ Assets/ScriptableObjects/Scripts/SquadStatSo.cs
@@ -12,7 +12,7 @@ namespace ScriptableObjects.Scripts
         public string squadStatName;
         
         [Header("초기값")]
-        public float initStatValue = 0;
+        public int initStatValue = 0;
         [Header("레벨 업 비용")]
         public int levelUpCost = 1;
 
@@ -21,6 +21,6 @@ namespace ScriptableObjects.Scripts
         [Header("스탯 증가량 타입")]
         public Data.Enum.IncreaseStatValueType increaseStatValueType;
         [Header("스탯 증가량")]
-        public float increaseStatValue;
+        public int increaseStatValue;
     }
 }
